pyproject.toml
[tool.poetry]
name = "easy-leasy"
version = "1.0.0"
description = "Easy Leasy is a time set manipulation library"
authors = ["Rémy Sanchez <remy.sanchez@hyperthese.net>"]
license = "WTFPL"
readme = "README.md"
packages = [
    { include = "easy_leasy", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "*"
coverage = "^7.6.4"
lark = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    # -- Core --
    "E",  # pycodestyle (errors)
    "W",  # pycodestyle (warnings)
    "F",  # Pyflakes
    "I",  # isort
    "C90",  # McCabe Complexity
    "D1",  # Mandatory docstrings
    "S",  # Bandit

    # -- Bonus --
    "A",  # flake8-bugbear
    "B",  # flake8-bugbear
    "DTZ",  # date/time zone
    "T10",  # debugger
    "EM",  # error messages
    "EXE",  # executable
    "G",  # logging
    "UP",  # pyupgrade
    "ASYNC",  # async
    "INP",  # PEP 420
    "T20",  # no print
    "PT",  # pytest
    "SLOT",  # slots
    "TCH",  # type checking blocks
    "PTH",  # Use pathlib
    "FIX",  # Don't allow not done stuff in code
    "ERA",  # Don't comment code
    "RUF",  # Ruff stuff
    "TID251",  # See below
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "D"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"requests".msg = "Use `httpx` instead, which has timeouts by default."

[tool.ruff.lint.pydocstyle]
convention = "numpy"
